pyproject.toml
[tool.poetry]
name = "social media integration feed service"
version = "0.1.0"
description = ""
authors = ["SaJH <wogur981208@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
cryptography = "^41.0.4"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.0"
djangorestframework-camel-case = "^1.4.2"
psycopg2-binary = "^2.9.9"
django-filter = "^23.3"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.9.1"
pre-commit = "^3.4.0"

[tool.black]
line-length = 140
target-version = ['py311']
force-exclude = 'migrations'

[tool.isort]
profile = "black"
filter_files = true
skip_glob = ["*/migrations/*", "config/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
